Revert "perf(api/view): small perf improvement, use a LinkedHashMap instead of Vec"
This reverts commit f775f3ee605b485f91e61b038cad697d34af8a8f.

Turns out this wouldn't work, views need to find their parents and add
itself to the parent's children. Since parents' children's type is
Vec<View>, using LinkedHashMap would be inconsistent, and is going to be
a lot more pain to do tree traversal.

It might be possible if the children type is Vec<View> (which I am now
thinking of doing), but I think I'm going to postpone this and maybe do
this performance improvement later.